Stop using TitleProxy
so that RealTimeSearchIndex will pick up the changes when CMS Titles are
updated - with the RealTimeSearchIndex it will check the model of the index
so when a Title is saved, the get_model() method needs to return Title for
the page to be re-indexed.

The downside of this way of fixing is that the verbose_name will refer to
titles rather than pages, but that seems the lesser of two evils.  Fixing
both properly might require small extensions to haystack itself.